perf: 整理了代码. 添加了注释
